Fixed cmake and build issues when using Visual Studio 2015

+++ modules/ocl/src/color.cpp
@@ -87,7 +87,7 @@ static void fromRGB_caller(const oclMat &src, oclMat &dst, int bidx, const std::
     if (!data2.empty())
         args.push_back( make_pair( sizeof(cl_mem) , (void *)&data2.data ));
 
-    size_t gt[3] = { dst.cols/pixels_per_work_item, dst.rows, 1 };
+    size_t gt[3] = { (size_t)dst.cols/pixels_per_work_item, (size_t)dst.rows, 1 };
 #ifdef ANDROID
     size_t lt[3] = { 16, 10, 1 };
 #else
@@ -122,7 +122,7 @@ static void toHSV_caller(const oclMat &src, oclMat &dst, int bidx, const std::st
     if (!data2.empty())
         args.push_back( make_pair( sizeof(cl_mem) , (void *)&data2.data ));
 
-   size_t gt[3] = { dst.cols, dst.rows, 1 };
+   size_t gt[3] = { (size_t)dst.cols, (size_t)dst.rows, 1 };
 #ifdef ANDROID
     size_t lt[3] = { 16, 10, 1 };
 #else
@@ -154,7 +154,7 @@ static void fromGray_caller(const oclMat &src, oclMat &dst, int bidx, const std:
     if (!data.empty())
         args.push_back( make_pair( sizeof(cl_mem) , (void *)&data.data ));
 
-    size_t gt[3] = { dst.cols, dst.rows, 1 };
+    size_t gt[3] = { (size_t)dst.cols, (size_t)dst.rows, 1 };
 #ifdef ANDROID
     size_t lt[3] = { 16, 10, 1 };
 #else
@@ -197,7 +197,7 @@ static void toRGB_caller(const oclMat &src, oclMat &dst, int bidx, const std::st
     if (!data.empty())
         args.push_back( make_pair( sizeof(cl_mem) , (void *)&data.data ));
 
-    size_t gt[3] = { dst.cols/pixels_per_work_item, dst.rows, 1 };
+    size_t gt[3] = { (size_t)dst.cols/pixels_per_work_item, (size_t)dst.rows, 1 };
 #ifdef ANDROID
     size_t lt[3] = { 16, 10, 1 };
 #else
@@ -229,7 +229,7 @@ static void toRGB_NV12_caller(const oclMat &src, oclMat &dst, int bidx, const st
     if (!data.empty())
         args.push_back( make_pair( sizeof(cl_mem) , (void *)&data.data ));
 
-    size_t gt[3] = {src.cols, src.rows, 1};
+    size_t gt[3] = {(size_t)src.cols, (size_t)src.rows, 1};
 #ifdef ANDROID
     size_t lt[3] = {16, 10, 1};
 #else
@@ -261,7 +261,7 @@ static void fromHSV_caller(const oclMat &src, oclMat &dst, int bidx, const std::
     if (!data.empty())
         args.push_back( make_pair( sizeof(cl_mem) , (void *)&data.data ));
 
-    size_t gt[3] = { dst.cols, dst.rows, 1 };
+    size_t gt[3] = { (size_t)dst.cols, (size_t)dst.rows, 1 };
 #ifdef ANDROID
     size_t lt[3] = { 16, 10, 1 };
 #else
@@ -288,7 +288,7 @@ static void RGB_caller(const oclMat &src, oclMat &dst, bool reverse)
     args.push_back( make_pair( sizeof(cl_int) , (void *)&src_offset ));
     args.push_back( make_pair( sizeof(cl_int) , (void *)&dst_offset ));
 
-    size_t gt[3] = { dst.cols, dst.rows, 1 };
+    size_t gt[3] = { (size_t)dst.cols, (size_t)dst.rows, 1 };
 #ifdef ANDROID
     size_t lt[3] = { 16, 10, 1 };
 #else
@@ -314,7 +314,7 @@ static void fromRGB5x5_caller(const oclMat &src, oclMat &dst, int bidx, int gree
     args.push_back( make_pair( sizeof(cl_int) , (void *)&src_offset ));
     args.push_back( make_pair( sizeof(cl_int) , (void *)&dst_offset ));
 
-    size_t gt[3] = { dst.cols, dst.rows, 1 };
+    size_t gt[3] = { (size_t)dst.cols, (size_t)dst.rows, 1 };
 #ifdef ANDROID
     size_t lt[3] = { 16, 10, 1 };
 #else
@@ -340,7 +340,7 @@ static void toRGB5x5_caller(const oclMat &src, oclMat &dst, int bidx, int greenb
     args.push_back( make_pair( sizeof(cl_int) , (void *)&src_offset ));
     args.push_back( make_pair( sizeof(cl_int) , (void *)&dst_offset ));
 
-    size_t gt[3] = { dst.cols, dst.rows, 1 };
+    size_t gt[3] = { (size_t)dst.cols, (size_t)dst.rows, 1 };
 #ifdef ANDROID
     size_t lt[3] = { 16, 10, 1 };
 #else
